week05/pyproject.toml
[tool.poetry]
name = "week05"
version = "0.1.0"
description = ""
authors = ["aalexren <chernitca@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.49.1"
grpcio-tools = "^1.49.1"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
